Do not allow the GPL
Julien pointed out that the inclusion of the GPL in the whitelist was an
error; which it was, I didn't notice this when I switched from my unix
scriptish approach to checking to using Jordan's code.

This removes the GPL, LGPL, and Artistic license; none of which we seem
to need.